third_party/WebKit/LayoutTests/fast/forms/form-radio-img-node-list-expected.txt

CONSOLE WARNING: The <keygen> element is deprecated and will be removed in M57, around March 2017. See https://www.chromestatus.com/features/5716060992962560 for more details.
Test RadioNodeLists returned by the HTMLFormElement named-getter.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Check that if no 'listed elements' match by name, img elements are picked instead.
PASS radioNodeList.length is 2
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList.length is 2
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList.length is 3
PASS radioNodeList[0] instanceof HTMLImageElement is true
PASS radioNodeList[1] instanceof HTMLImageElement is true
PASS radioNodeList[2] instanceof HTMLImageElement is true
PASS successfullyParsed is true

TEST COMPLETE

                    
